Things are happening

+++ Models/Drinks.cs
@@ -2,7 +2,7 @@ namespace BurgerShack.Models
 {
   public class Drinks : MenuItem
   {
-    public Drinks(string title, int price, string desc) : base(title, price, desc)
+    public Drinks(string title, double price, string desc) : base(title, price, desc)
     {
     }
   }
