+++ Models/Entree.cs
@@ -2,7 +2,7 @@ namespace BurgerShack.Models
 {
   public class Entree : MenuItem
   {
-    public Entree(string title, int price, string desc) : base(title, price, desc)
+    public Entree(string title, double price, string desc) : base(title, price, desc)
     {
     }
   }
